Add comprehensive documentation
Documentation for the Midoridan project, a Discord bot program, is detailed in an extensive README. Included are sections dedicated to its features, code structure, standards, prerequisites, installation and usage procedures, configuration, and development practices such as linting and formatting. The README also outlines how to contribute to the project and information on its GPL-3.0 License.