Added verbose logging & fixed experimental warning
In this update, a verbose configuration option `VERBOSE` was introduced in .env-sample file and utilized in src/utils.ts for extensive logging. Type expectation was added in src/config.ts for better type safety.
Also, a code block was added to src/utils.ts to suppress 'ExperimentalWarning' on console to make logs cleaner. The changes will help in debugging and offer better code clarity.