Updated API endpoints and image processing parameters
The URL for TEXT_TO_IMAGE_API and IMAGE_TO_IMAGE_API in the .env-sample file has been modified. The new links now point to the 'stable-diffusion-xl-1024-v1-0' version. Additionally, the 'createRequestBody' function in the utils.ts file has been altered to change the image processing parameters. The number of steps is now 40, and the width and height values have been changed from 512 to 1024.