Added automatic directory creation for image output
The saveImages function now checks for and automatically creates a designated output directory if it doesn't exist. This was achieved by importing the 'path' module and using it to enhance the file path generation. A constant was set for the 'outputFolder', instead of hardcoding the string in the path, in case we need to change the folder at a later point. All related file paths in the function were updated accordingly.